feat: add lib.rs for benchmark
