+++ Cargo.toml
@@ -60,3 +60,4 @@ regex = "1.7.0"
 
 [dev-dependencies]
 httpmock = "0.6.6"
+criterion = "0.4"
